workshop/requirements.txt
# AP2 Workshop Requirements

# Google Agent Development Kit with A2A support
google-adk[a2a]>=1.0.0

# Pydantic for data models
pydantic>=2.0.0

# Environment variable management
python-dotenv>=1.0.0
